+++ sqlitecluster/SQLite.h
@@ -454,47 +454,3 @@ class SQLite {
     const string _synchronous;
     int64_t _mmapSizeGB;
 };
-
-class SQLitePool {
-  public:
-    // Create a pool of DB handles.
-    SQLitePool(size_t maxDBs, const string& filename, int cacheSize, bool enableFullCheckpoints, int maxJournalSize,
-               int createJournalTables, const string& synchronous = "", int64_t mmapSizeGB = 0, bool pageLoggingEnabled = false);
-    ~SQLitePool();
-
-    // Get the base object (the first one created, which uses the `journal` table). Waits until this handle is
-    // available if it isn't.
-    SQLite& getBase();
-
-    // Get any object except the base. Will wait for an available handle if there are already maxDBs.
-    SQLite& get();
-
-    // Return an object to the pool.
-    void returnToPool(SQLite& object);
-
-  private:
-    // Synchronization variables.
-    mutex _sync;
-    condition_variable _wait;
-
-    size_t _maxDBs;
-    // Pointers to all of our objects.
-    SQLite* _baseDB;
-    set<SQLite*> _availableHandles;
-    set<SQLite*> _inUseHandles;
-};
-
-class SQLitePoolScopedGet {
-  public:
-    SQLitePoolScopedGet(SQLitePool& pool) : _pool(pool), _db(_pool.get()) {} 
-    ~SQLitePoolScopedGet() {
-        _pool.returnToPool(_db);
-    }
-    SQLite& db() {
-        return _db;
-    }
-
-  private:
-    SQLitePool& _pool;
-    SQLite& _db;
-};
